Fleshing out the README with TODOs and plans
Roughing out how I'm going to change the division of responsibility
between the parser and compiler.  At present the compiler does some very
simple environment stuff for function application but that should move
to the parsing stage.  The compiler should be handed an AST where all of
the resolution stuff is already complete and no further error checking
needs to be done.