Function application using AST record
Also removed the "nesting" tests, unifying under expression tests for
expressions nested in parens.  Experimenting with the parser returning
an error as a success in the case of applying arguments to anything
other than a symbol.  This is short sighted at the moment since one
should be able to apply arguments to a nested expression that yields a
function.