let-polymorphism coming together
Working inferencer for integers and floats where a function bound inside
a function with let is actually polymorphic and not bound to the first
types it's applied to.

Code is a bit of a messy horror-show at the moment with environment
dumping and io:format calls everywhere but wanted to get this committed
while it's working.