Naive first pass at single module typing
Basically just a fold over the defined functions and environment that's
passed in.  Unsolved issues (hence this being naive) are the inability
to refer to a function's type where the function called is defined later
than the function being typed and neither checking arity correctness of
exported functions nor binding them in the environment for later
modules.  This latter issue might best be solved by the full program
typing function.