Type modules with forward references
Typing of inter-module calls is not complete in this commit but I wanted
to get this forward reference work committed separately.  By "forward
reference" I mean when a function refers to a top level function that
occurs later in the module's definition and thus is not yet bound in the
environment.