Starting work on ADT type instantiation
Converting a #mlfe_type{} record to an instance of an #adt{} record for
use in unification.  Some dialyzer errors here, just checking in
progress so the changes are a bit more discrete/evident.

I'm creating all ADT type variables at level 0 right now and have
included comments to that effect.  My thinking is that types are always
declared at the top level but it may make sense to actually carry the
typing level here as in other areas.  I suspect I may run into issues
later when generalization is applied to functions that accept
polymorphic ADTs.