Add a test to ensure ADT unify preserves var cells
The single test actually shows that the typer appears to correctly
preserve the reference cells in instantiated polymorphic ADTs *without*
an explicit deep copy.  This is a surprising result to me, I thought
that the cells would not be protected and that the first function that
unifies with them would set a type that would prohibit unification with
other.  I may need to investigate further.