Compiling forward references now works
It was failing because when a function wasn't in the common erlang code
generator's environment (see mlfe_code_gen.erl), the generator didn't
know the arity of the function being called.  If it didn't know the
arity, the compiler handling core erlang couldn't look it up properly
and compilation failed.