More accurate line# for clause unify type errors
Previously when clauses in the clauses for a match or receive could not
unify, the line reported for the error was that of the actual occurrence
of "match" or "receive".  This is only correct when the type of the
match expression does not unify with the patterns of all clauses.
unify_clauses/3 now uses the line from each clause AST node when
attempting unification with the current typing of the set of clauses.