Closes #25
Changes to copying receivers so that the receive type and the receive's
expression type use the same reference cells for the same type
variables.  A bit rough around the edges at the moment, we might have
some issues with complex receive types that I haven't tested yet.

Fails to dialyze on 19.0 but passes on 18.3 and 19.1 which I'm happy
enough with.