Basic record compilation to maps
Maps are far easier to compile to for now so I went this direction in
order to deliver something working sooner.  I still think it's valuable
to compile to tuples as well and perhaps provide a compiler option to
switch between the two depending on what the user prefers.

New bug:  multiple occurences of `_` in a pattern are causing a
compilation error in the Core Erlang compiler, just need to add some
renaming in the AST generation step.

Still missing:
- field accessors (e.g. f.x)
- field punning (e.g. {x, y} instead of {x=x1, y=y1} in a match)
- record updates.  Not sure what I want to do here syntax-wise yet.
Single field updates are relatively simple, multiples less so perhaps.