Basics of exporting all arities for a function
This was suggested in #63.  It's not actually part of the AST generation
phase yet but it made sense to handle this at the same time as function
imports for #62.  I'm going to decouple the basic parsing pass from the
renaming and rewriting pass of the parser since rewriting the names of
imported functions to inter-module calls will require the other modules
to already be parsed and accessible.