Single multiple module AST generation function
make_modules/1 takes a list of source code strings to parse and rewrite
as proper modules with imports and exports expanded.  Expanded exports
now work so we can do things like `export foo` instead of `export foo/1,
foo/2`.  Rewriting functions not defined in the module to inter-module
calls using imports is not yet done but I think because of the
increasing amount of changes it's worth integrating now, if only to get
the export changes in.

Fixes #63