Perform exhaustiveness checking of top level functions.
Provides an option `warn_exhaustiveness` that will cause warnings to be
printed in case a top level function does not cover all the possible
patterns for its argument types.

Note that the check will be run regardless of whether the option is set
or not. The option only affects printing of warnings messages. This is
to ensure the checker evolves with the rest of the code.

No exhaustiveness checks are provided for maps. A true exhaustiveness
check would check that all combinations of presence/absence of the
possible keys for the key type match all the possible combinations of
the value type. This is clearly not of any practical value though.

A more pragmatic exhaustiveness check for maps would be to check value
coverage for all provided keys. However, this requires looking at the
provided patterns to find the covering patterns, which the current
algorithm does not support.