Concrete types in types as type members
Not applied to module-qualified types yet.  This lets the following
actually work:

    type opt 'a = Some 'a | None
    type int_opt = opt int

Up until this commit any attempt to use `int_opt` would cause a
unification failure because we weren't capturing `int` as an actual
parameter and associated variable name.  Now for each parameter that is
a concrete type we synthesize a variable name for it.