Pattern matching for unary lambdas
Fixes #138

A lambda that has a single argument can just use clauses instead of
explicitly using a match ... with.  I've enabled guards for functions as
well although they're only supported in the syntax for this new kind of
lambda so far.  We probably want to limit them to type checks and maybe
strict equality as per discussion with @danabr in issue #95.