Problem: var lookup fails in exhaustiveness checks
ADTs that referred to other ADTs (e.g. a type that uses the option type)
that:

- use different variable names
- use concrete types, leading to synthetic variable names

were causing crashes in the exhaustiveness checker.  Caught this when
using options inside of records, the same issue occurred with lists and
maps as well.  Now the exhaustiveness checker substitutes the original
variable names before checking each type variable for coverage.