Fixes #176, nested user ADTs failed to unify
There are two paths for instantiating types, one for when they're
defined at top-level and another for when we're instantiating a type
constructor's argument for type arrows.  The latter is a little
different in that it handles reusing and replacement of variables a
little differently.  It's entirely possible that we should unify these
two to cut down on errors.

The basic problem was that when instantiating constructor arguments we
were recursing correctly when instantiating a types members but _not_
when instantiating type variables that were bound to a particular type.
That meant that this would be OK:

    type opt 'a = Some 'a | None
    type w 'z = W opt (opt 'z)

But this would not:

    type opt 'a = Some 'a | None
    type w = W opt (opt int)