Build with last Alpaca and then rebuild with master
To handle AST changes this builds Alpaca with the current official
release, makes a _test_ release and then rebuilds Alpaca and runs
tests with that build of master.  I think this will be acceptable for
now as the previously tested typer is still run against all of the
main code and then we additionally validate it with itself.