AST and Code generation symbol -> label changes
This had the side-effect of making some inter-module checks more
strict earlier in the AST generation which I think is not a bad
thing.  It basically meant that some code generation tests which
separately compiled modules need to now include them all up front.
This additionally made one of the stacktrace tests redundant, changes
coming in a later commit.