Symbol becomes "label", added "qualified labels" (#267)
Label is a bit more accurate than "symbol" for this sort of node in terms of its function in the AST.

"Qualified labels" are those which are namespaced, e.g. `module.function` or `record.field` (though the latter doesn't work yet.)

This was a fairly invasive change that also impacted how module names are handled but I think it's a relatively big step towards a lot more consistency and clarity in the AST.

* Renaming "symbol" to "label", adding "qualified labels"

"Label" is a more appropriate name for how these things are used since
they're quite literally labelling variables, modules, etc.

Qualified labels are those that are prefixed with some sort of
namespace, like a module or record, e.g. `lists.keyfind`.

* Parser and scanner changes for symbol -> label

Lots of this is trivial renaming but there are some more substantial
changes in how module names are handled now.  Since they're initially
captured as labels which use binaries for names, I carried this binary
format through elsewhere.  Following commits will show the more
invasive changes this required.

* AST and Code generation symbol -> label changes

This had the side-effect of making some inter-module checks more
strict earlier in the AST generation which I think is not a bad
thing.  It basically meant that some code generation tests which
separately compiled modules need to now include them all up front.
This additionally made one of the stacktrace tests redundant, changes
coming in a later commit.

* Typer changes for symbol -> label

Lots of atom -> binary things which required a lot of
spelunking (missed having static typing in this pass.)  Took the
opportunity to document a few things more, and found a private types
bug as well which will need fixing but I might defer to when we get
into abstract types.

* Exhaustiveness and printer symbol -> label changes

Nothing particularly drastic here although both modules' tests helped
turn up some earlier silly mistakes.

* Final symbol -> label changes

Added a version of `compile/2` to help tests that have multiple
modules in memory rather than on-disk.

* alpaca.hrl to allow general use of `compiled_module` record